change structure

+++ Uni/OOP/HW 1/Task 2/System.h
@@ -25,14 +25,14 @@ public:
     System();
     System(size_t size);
     ~System();
-    void addFile(const char* name, size_t editDay, size_t createMonth, size_t createYear,
-        size_t creatHours, size_t createMins, size_t createSeconds, const char* permissions);
+    void addFile(const char* name, unsigned editDay, unsigned createMonth, unsigned createYear,
+        unsigned creatHours, unsigned createMins, unsigned createSeconds, const char* permissions);
 
-    void editFile(const char* name, const char* content, size_t editDay, size_t editMonth, size_t editYear,
-                          size_t editHours, size_t editMins, size_t editSeconds, char user);
+    void editFile(const char* name, const char* content, unsigned editDay, unsigned editMonth, unsigned editYear,
+                          unsigned editHours, unsigned editMins, unsigned editSeconds, char user);
 
-    void addInFile(const char* name, const char* content, size_t editDay, size_t editMonth, size_t editYear,
-                          size_t editHours, size_t editMins, size_t editSeconds, char user);
+    void addInFile(const char* name, const char* content, unsigned editDay, unsigned editMonth, unsigned editYear,
+                          unsigned editHours, unsigned editMins, unsigned editSeconds, char user);
 
     void deleteFile(const char* name);
     void editRights(const char* name, char group, char right);
